Modify OnBoarding image with circle around it
