--- app/src/main/java/android/tvz/hr/fitnessclienttracker/onboarding/ThirdScreenFragment.kt
@@ -1,23 +1,0 @@
-package android.tvz.hr.fitnessclienttracker.onboarding
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.material.Text
-import androidx.compose.ui.platform.ComposeView
-import androidx.fragment.app.Fragment
-
-class ThirdScreenFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        return ComposeView(requireContext()).apply {
-            setContent {
-                Text(text = "Treći")
-            }
-        }
-    }
-}
