+++ app/src/main/java/android/tvz/hr/fitnessclienttracker/onboarding/ViewPagerFragment.kt
@@ -6,7 +6,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.tvz.hr.fitnessclienttracker.R
-import androidx.compose.material.Text
+import android.tvz.hr.fitnessclienttracker.onboarding.screens.FirstScreenFragment
+import android.tvz.hr.fitnessclienttracker.onboarding.screens.SecondScreenFragment
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.viewpager2.widget.ViewPager2
@@ -26,14 +27,14 @@ class ViewPagerFragment : Fragment() {
                     FirstScreenFragment(),
                     SecondScreenFragment()
                 )
-
                 val adapter = ViewPagerAdapter(
                     fragmentList,
                     requireActivity().supportFragmentManager,
                     lifecycle
                 )
                 AndroidView(factory = { context ->
-                    val view = LayoutInflater.from(context).inflate(R.layout.fragment_view_pager, null, false)
+                    val view = LayoutInflater.from(context)
+                        .inflate(R.layout.fragment_view_pager, null, false)
                     val viewPager = view.findViewById<ViewPager2>(R.id.viewPager)
                     viewPager.adapter = adapter
                     view
