+++ app/src/main/java/android/tvz/hr/fitnessclienttracker/onboarding/screens/ThirdScreenFragment.kt
@@ -1,4 +1,4 @@
-package android.tvz.hr.fitnessclienttracker.onboarding
+package android.tvz.hr.fitnessclienttracker.onboarding.screens
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,7 +8,7 @@ import androidx.compose.material.Text
 import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 
-class FirstScreenFragment : Fragment() {
+class ThirdScreenFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -16,7 +16,7 @@ class FirstScreenFragment : Fragment() {
     ): View {
         return ComposeView(requireContext()).apply {
             setContent {
-                Text(text = "Prvi")
+                Text(text = "Treći")
             }
         }
     }
